apps/update-rag/requirements.txt

azure-identity
azure-storage-blob
azure-cosmos
azure-search-documents
requests
tqdm
python-dotenv
openai